Move MockDistServer into rust_install so other test cases can use it
