+++ rust-install/src/lib.rs
@@ -17,6 +17,9 @@ extern crate ole32;
 extern crate rust_manifest;
 extern crate tempdir;
 extern crate walkdir;
+extern crate toml;
+extern crate flate2;
+extern crate tar;
 
 pub use errors::*;
 pub use install::*;
